tool/backend/requirements.txt
Django
djangorestframework
graphene-django
django-environ
psycopg2
django-cors-headers
pylint-django